User signin using username and not email #13 with minor refactoring of code
